Fix issues with circular imports introduced by type annotations.
